Moved login Post processor to UsernameAuthenticationScheme.
